Supply discovery options for logging
Summary:
When going through MyRouter the database disovery options are added to the samples in the xstream pipeline when derived from the MR3 encoded database name (https://fburl.com/code/fhnh9n5v).

In Dvaar we don't use the MR3 encoded database name while connecting so we don't have any of these fields currently filled in.

In order to handle this, let's put the MR3 encoded database name into the FBConnectionContext and then during logging we can decode it and fill in those fields.

Differential Revision: D74494892

fbshipit-source-id: b6e3c34db80c52f46ff5fc9f89dba68796829304