Fix race condition
Summary:
We had a race condition for operations that never got started (i.e. `run()` was never called).  They could be cleaned up the normal way or via the `drain()` function on shutdown.  Each method was not acting aware of the other so whichever happened first could cause a crash in the other.

Add the checks to keep all the correct data structures in line with each other so that race condition can no longer occur.

Differential Revision: D75324203

fbshipit-source-id: 6c4d1b5f8d7bbe52d0048230fee194442b865585