Allow for moving PoolOptions
Summary:
In D75182675 we allowed the connection pool to dynamically update the per-key connection limit.  The thought was to make the field in the PoolOptions atomic to protect it.

However, this causes the PoolOption to not be default movable and upon further reflection, the PoolOptions struct itself should not be forcing the multi-threaded protection.

Change the PoolOptions back to not have an atomic field.  Then only the connection pool's copy of the PoolOptions is protected against threaded changes via `folly::Synchronized`.  This is more clean and clear and allows for other fields to be changed in the future.

Reviewed By: fadimounir

Differential Revision: D79654180

fbshipit-source-id: 78d1bea2816d8072401c9c04a4edb8509896e291