Pass logging functions via query operation
Summary:
The original implementation of adding query logging functions (functions that don't do the work to resolve the value for the specified key unless they are needed for logging) (done in D81636679) was flawed.  It attempted to use the connection context to store the functions.  This worked _most_ of the time, but could cause a TSAN exception.

The main issue is that when we start a query we overwrite the connection context's logging functions (taken from the query options), but logging for the connection's success occurs _after_ the connection is available to the user again.  This causes a TSAN error if the logging for the successful connection reads the logging functions right as the query attempts to overwrite it.

The correct way to do this is to pass the logging function information through the query operation instead of through the connection context.

This diff moves that all around to achieve the correct result.

Reviewed By: fadimounir

Differential Revision: D87899124

fbshipit-source-id: af2054400d8b0949ede1880bb67d23065d9b758a